convert some functions to ts
